+++ src/templates/index.ts
@@ -1,17 +1,17 @@
 import { TemplatesRepository } from './infra/repositories/TemplatesRepository'
 import { CreateTemplate } from './domain/useCases/CreateTemplate'
 import { DeleteTemplate } from './domain/useCases/DeleteTemplate'
-import { GetDatasetTemplates } from './domain/useCases/GetDatasetTemplates'
+import { GetTemplatesByCollectionId } from './domain/useCases/GetTemplatesByCollectionId'
 import { GetTemplate } from './domain/useCases/GetTemplate'
 
 const templatesRepository = new TemplatesRepository()
 
 const createTemplate = new CreateTemplate(templatesRepository)
 const deleteTemplate = new DeleteTemplate(templatesRepository)
-const getDatasetTemplates = new GetDatasetTemplates(templatesRepository)
+const getTemplatesByCollectionId = new GetTemplatesByCollectionId(templatesRepository)
 const getTemplate = new GetTemplate(templatesRepository)
 
-export { createTemplate, deleteTemplate, getDatasetTemplates, getTemplate }
+export { createTemplate, deleteTemplate, getTemplatesByCollectionId, getTemplate }
 export {
   CreateTemplateDTO,
   TemplateFieldDTO,