+++ test/functional/templates/createTemplate.test.ts
@@ -1,7 +1,7 @@
 import { ApiConfig } from '../../../src'
 import { DataverseApiAuthMechanism } from '../../../src/core/infra/repositories/ApiConfig'
 import { TestConstants } from '../../testHelpers/TestConstants'
-import { createTemplate, getDatasetTemplates } from '../../../src/templates'
+import { createTemplate, getTemplatesByCollectionId } from '../../../src/templates'
 import { CreateTemplateDTO } from '../../../src/templates/domain/dtos/CreateTemplateDTO'
 import { MetadataFieldTypeClass } from '../../../src/metadataBlocks/domain/models/MetadataBlock'
 import { deleteDatasetTemplateViaApi } from '../../testHelpers/datasets/datasetTemplatesHelper'
@@ -48,7 +48,7 @@ describe('CreateTemplate.execute', () => {
       ]
     }
     await createTemplate.execute(templateDto)
-    const templates = await getDatasetTemplates.execute(':root')
+    const templates = await getTemplatesByCollectionId.execute(':root')
 
     expect(templates[templates.length - 1].name).toBe(templateDto.name)
     expect(templates[templates.length - 1].isDefault).toBe(templateDto.isDefault)